+++ noodles-cram/src/data_container/compression_header/preservation_map/builder.rs
@@ -1,6 +1,6 @@
 use super::{substitution_matrix, tag_ids_dictionary, PreservationMap};
 
-use crate::Record;
+use crate::{writer::Options, Record};
 
 #[derive(Debug)]
 pub struct Builder {
@@ -12,6 +12,12 @@ pub struct Builder {
 }
 
 impl Builder {
+    pub fn apply_options(&mut self, options: &Options) {
+        self.read_names_included = options.preserve_read_names;
+        self.ap_data_series_delta = options.encode_alignment_start_positions_as_deltas;
+        self.reference_required = options.require_reference_sequence;
+    }
+
     pub fn update(&mut self, reference_sequence: &[u8], record: &Record) {
         self.substitution_matrix_builder
             .update(reference_sequence, record);